Finally resolved problem with combo of Ubuntu, virtualenv and local test discovery
The problem was that the test doscovery process was doscovering tests via the symlink
paths in (env)/local/lib, and detecting a mismatch with the real path of the test modules.

The problem is resolved by using os.path.realpath when calculating the path for the
annalist_root module.  Also added some extra test runner parameters to stop Django
searching all installed packages for potential test cases.