Fix (benign but ugly) bug causing 404 errors when an entity is renamed.
The problem occurred when the entity was edited in a form invoked from a view of the same entity.
On completion of the edit, returning to the view would fail because the URI had changed.  The fix
involves updating the continuation URIs that are used wen the editing form is dismissed to take
account of the renaming.