Added type cache to Collection class to reverse slow-down introduced by subtype search logic. Collection now provides direct access to types by type id and by type URI; this speeds up the test suite by about 20%.
Changed EntityTypeInfo to use collection class to access type data (hence the cache).

Change cache logic to make it more robust in the face of changes.

Note that the cache is refreshed for each HTTP request, in part to reduce the risk of
problems caused by stale data.  It is possible for changes to the type data to bypass
the collection class cache.

All tests pass.